feat(gimmighouls): el gacha baja a 0.01 por dólar, y el contador pasa a decimal
Bajar el ratio a secas no bajaba la recompensa: la borraba. El contador era entero y
`award_gimmighouls` redondeaba cada premio por separado, así que un sobre de 50 $ valía
0.5 y `round(0.5)` es 0. Las dos máquinas MÁS jugadas, las de 25 y 50 $, habrían dejado
de pagar absolutamente nada mientras la ayuda seguía prometiendo 0.01 por dólar.

El contador es decimal. No hay redondeo en ningún premio, así que tampoco se paga de más:
un sobre de 150 $ da 1.5, no 2. Lo único que se recorta son los seis decimales que hacen
falta para no guardar la basura de la coma flotante binaria (100 $ con un boost del 10%
daban 110.00000000000001).

Las bases de datos que ya existen NO necesitan migración: SQLite tipa por afinidad y
guarda 0.5 tal cual aunque la columna esté declarada INTEGER, porque solo convierte a
entero cuando no se pierde nada. El FLOAT del modelo dice la intención y es lo que haría
falta si algún día esto no fuera SQLite.

En pantalla, un único formateador. Lo pintaban cuatro sitios con dos formateadores
distintos copiados (cabecera de escritorio, cabecera de móvil, ranking ancho y estrecho),
y con decimales bastaba que uno se quedara atrás para enseñar un número que no se parece
a los otros tres. Dos decimales: hacen visible medio punto sin sacar el 36.465 de un boost.

Comprobado con las máquinas reales: 25 $ → 0.25, 50 $ → 0.5, 250 $ → 2.5, y diez tiradas
seguidas en la de 50 $ suman los 5 puntos que tocan.

+++ backend/app/models.py
@@ -15,7 +15,9 @@ class User(Base):
     alias: Mapped[Optional[str]] = mapped_column(String, nullable=True)
     elo: Mapped[int] = mapped_column(Integer, default=1200)
     games_played: Mapped[int] = mapped_column(Integer, default=0)
-    gimmighouls: Mapped[int] = mapped_column(Integer, default=0)
+    #: Decimal a propósito: el gacha paga 0.01 por dólar, así que los premios son fracciones de punto
+    #: y un contador entero los redondearía a cero. Ver `award_gimmighouls`.
+    gimmighouls: Mapped[float] = mapped_column(Float, default=0.0)
     referred_by: Mapped[Optional[str]] = mapped_column(String, nullable=True)  # ReferralCode.code
     withdraw_address: Mapped[Optional[str]] = mapped_column(String, nullable=True)  # USDC payout destination
     emote_slots: Mapped[Optional[str]] = mapped_column(String, nullable=True)  # JSON list of up to 3 quick-access emote codes
@@ -45,7 +47,7 @@ class ReferralCode(Base):
     # paga lo mismo.
     rake_share_pct: Mapped[float] = mapped_column(Float, default=0.25)
     owner_wallet: Mapped[Optional[str]] = mapped_column(String, nullable=True)  # User to credit the cut to
-    earned: Mapped[int] = mapped_column(Integer, default=0)  # fallback tally when no owner_wallet
+    earned: Mapped[float] = mapped_column(Float, default=0.0)  # fallback tally when no owner_wallet
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=_now)
